+++ frontend/src/components/Geral/PaginaNotFound/NotFound.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default function NotFound(){
+    return(
+        <React.Fragment>
+            <div className="container">
+                <div className="row py-5">
+                    <div className="col-12">
+                        <p>Página não encontrada.</p>
+                    </div>
+                </div>
+            </div>
+        </React.Fragment>
+    )
+}