+++ frontend/src/components/Pacientes/Navbar/Navbar.jsx
@@ -7,7 +7,9 @@ import Home from '../../../pages/Pacientes/Home/';
 import Perfil from '../../../pages/Pacientes/Perfil';
 import AuthPaciente from '../../../services/AuthPaciente';
 import Loading from '../../Geral/Loading';
-import Consultar from '../../Geral/Consulta/Consultar';
+import Consultas from '../../Geral/Consulta/Consultar';
+import Sala from '../../Geral/Consulta/Sala';
+import NotFound from '../../Geral/PaginaNotFound/NotFound';
 
 const Navbar = () => {
 
@@ -41,30 +43,33 @@ const Navbar = () => {
                 (
                     <>
                         <div className="container-fluid navBarToggle">
-                            <div className="row">
-                                <div className="navbar">
-                                    <div className="col-12">
-                                        <ul className="nav-menu-items">
-                                            {NavbarElements.map((item, index) => {
-                                                return(
-                                                    <li key={index} className={item.cName}>
-                                                        <Link to={item.link} onClick={fetchDadosPaciente}>
-                                                            <span className="iconeSpan">{item.icon}</span>
-                                                            <span className="nav-titulo">{item.titulo}</span>
-                                                        </Link>
-                                                    </li>
-                                                )
-                                            })}
-                                        </ul>
+                            <div className="container">
+                                <div className="row mx-auto">
+                                    <div className="navbar">
+                                        <div className="col-12 text-center">
+                                            <ul className="nav-menu-items">
+                                                {NavbarElements.map((item, index) => {
+                                                    return(
+                                                        <li key={index} className={item.cName}>
+                                                            <Link to={item.link} onClick={fetchDadosPaciente}>
+                                                                <span className="iconeSpan">{item.icon}</span>
+                                                                <span className="nav-titulo">{item.titulo}</span>
+                                                            </Link>
+                                                        </li>
+                                                    )
+                                                })}
+                                            </ul>
+                                        </div>
                                     </div>
                                 </div>
                             </div>
                         </div>
 
                         <Routes>
-                            <Route path="/" element={<Home paciente={dados_paciente}/>} />
+                            <Route exact path="/" element={<Home paciente={dados_paciente}/>} />
                             <Route path="/perfil" element={<Perfil paciente={dados_paciente}/>} />
-                            <Route path="/consultar-agora" element={<Consultar usuario={dados_paciente}/>} />
+                            <Route path="/consultas" element={<Consultas usuario={dados_paciente}/>} />
+                            <Route exact path="/consultas/sala/:id" element={<Sala usuario={dados_paciente}/>} />
                         </Routes>
                     </>
                 )