+++ frontend/src/services/MediaHandler.js
@@ -1,13 +1,16 @@
 export default class MediaHandler {
     getPermissoes(){
         return new Promise((resolve, rejected) => {
-            navigator.mediaDevices.getUserMedia({video: true, audio: true})
-                .then((stream) => {
-                    resolve(stream);
-                })
-                .catch(erro => {
+            navigator.mediaDevices.getUserMedia({
+                video: true, 
+                audio: true
+            }).then((stream) => {
+            
+                resolve(stream);
+            
+            }).catch(erro => {
                     throw new Error(`Impossível iniciar a transmissão de vídeo: ${erro}`)
-                })
-        })
+            });
+        });
     }
 }